+++ scripts/embedding/xsb/xsb_logtalk_xwam.sh
@@ -5,7 +5,7 @@
 ##   This script creates a XSB logtalk.xwam file
 ##   with the Logtalk compiler and runtime
 ## 
-##   Last updated on April 16, 2018
+##   Last updated on May 16, 2018
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -48,16 +48,16 @@ if ! [ "$LOGTALKHOME" ]; then
 		LOGTALKHOME="$( cd "$( dirname "$0" )" && pwd )/.."
 		echo "... using Logtalk installation found at $( cd "$( dirname "$0" )" && pwd )/.."
 	else
-		echo "... unable to locate Logtalk installation directory!"
+		echo "... unable to locate Logtalk installation directory!" >&2
 		echo
 		exit 1
 	fi
 	echo
 	export LOGTALKHOME=$LOGTALKHOME
 elif ! [ -d "$LOGTALKHOME" ]; then
-	echo "The environment variable LOGTALKHOME points to a non-existing directory!"
-	echo "Its current value is: $LOGTALKHOME"
-	echo "The variable must be set to your Logtalk installation directory!"
+	echo "The environment variable LOGTALKHOME points to a non-existing directory!" >&2
+	echo "Its current value is: $LOGTALKHOME" >&2
+	echo "The variable must be set to your Logtalk installation directory!" >&2
 	echo
 	exit 1
 fi
@@ -125,7 +125,7 @@ directory="$HOME/collect"
 paths="$LOGTALKHOME/paths/paths_core.pl"
 
 print_version() {
-	echo "$(basename "$0") 0.9"
+	echo "$(basename "$0") 0.10"
 	exit 0
 }
 
@@ -149,7 +149,6 @@ usage_help()
 	echo "  -v print version of $(basename "$0")"
 	echo "  -h help"
 	echo
-	exit 0
 }
 
 while getopts "cd:p:h:l:s:vh" option
@@ -158,12 +157,11 @@ do
 		c) compile="true";;
 		d) d_arg="$OPTARG";;
 		p) p_arg="$OPTARG";;
-		h) h_arg="$OPTARG";;
-		l) l_arg="$OPTARG";;
 		s) s_arg="$OPTARG";;
+		l) l_arg="$OPTARG";;
 		v) print_version;;
-		h) usage_help;;
-		*) usage_help;;
+		h) usage_help; exit;;
+		*) usage_help; exit;;
 	esac
 done
 
@@ -175,16 +173,7 @@ if [ "$p_arg" != "" ] ; then
 	if [ -f "$p_arg" ] ; then
 		paths="$p_arg"
 	else
-		echo "The $p_arg library paths file does not exist!"
-		exit 1
-	fi
-fi
-
-if [ "$h_arg" != "" ] ; then
-	if [ -f "$h_arg" ] ; then
-		hooks="$h_arg"
-	else
-		echo "The $h_arg hooks file does not exist!"
+		echo "The $p_arg library paths file does not exist!" >&2
 		exit 1
 	fi
 fi
@@ -193,7 +182,7 @@ if [ "$l_arg" != "" ] ; then
 	if [ -f "$l_arg" ] ; then
 		loader="$l_arg"
 	else
-		echo "The $l_arg loader file does not exist!"
+		echo "The $l_arg loader file does not exist!" >&2
 		exit 1
 	fi
 else
@@ -204,7 +193,7 @@ if [ "$s_arg" != "" ] ; then
 	if [ -f "$s_arg" ] ; then
 		settings="$s_arg"
 	else
-		echo "The $s_arg settings file does not exist!"
+		echo "The $s_arg settings file does not exist!" >&2
 		exit 1
 	fi
 else
@@ -212,7 +201,7 @@ else
 fi
 
 mkdir -p "$directory"
-cd "$directory"
+cd "$directory" || exit 1
 
 operating_system=$(uname -s)
 
@@ -271,18 +260,18 @@ fi
 
 xsb -e "compile(logtalk),halt."
 
-rm *.pl
-rm *.P
-rm *_pl.xwam
-rm *_lgt.xwam
+rm ./*.pl
+rm ./*.P
+rm ./*_pl.xwam
+rm ./*_lgt.xwam
 
 if [ "$loader" != "" ] ; then
 	mkdir -p "$directory/application"
-	cd "$directory/application"
+	cd "$directory/application" || exit 1
 	xsblgt$extension -e "set_logtalk_flag(clean,off),set_logtalk_flag(scratch_directory,'$directory/application'),logtalk_load('$loader'),halt."
-	cat $(ls -rt $directory/application/*.P) > application.P
+	cat $(ls -rt "$directory/application"/*.P) > application.P
 	xsblgt$extension -e "compile(application),halt."
 	mv application.xwam ..
-	rm *.P
-	rm *.xwam
+	rm ./*.P
+	rm ./*.xwam
 fi