+++ tools/lgtdoc/xml/lgt2html.sh
@@ -3,7 +3,7 @@
 #############################################################################
 ## 
 ##   XML documenting files to (X)HTML conversion script 
-##   Last updated on February 8, 2018
+##   Last updated on May 16, 2018
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -46,15 +46,15 @@ if ! [ "$LOGTALKHOME" ]; then
 		LOGTALKHOME="$( cd "$( dirname "$0" )" && pwd )/.."
 		echo "... using Logtalk installation found at $( cd "$( dirname "$0" )" && pwd )/.."
 	else
-		echo "... unable to locate Logtalk installation directory!"
+		echo "... unable to locate Logtalk installation directory!" >&2
 		echo
 		exit 1
 	fi
 	echo
 elif ! [ -d "$LOGTALKHOME" ]; then
-	echo "The environment variable LOGTALKHOME points to a non-existing directory!"
-	echo "Its current value is: $LOGTALKHOME"
-	echo "The variable must be set to your Logtalk installation directory!"
+	echo "The environment variable LOGTALKHOME points to a non-existing directory!" >&2
+	echo "Its current value is: $LOGTALKHOME" >&2
+	echo "The variable must be set to your Logtalk installation directory!" >&2
 	echo
 	exit 1
 fi
@@ -103,8 +103,8 @@ usage_help()
 	echo "current directory to XHTML or HTML files"
 	echo
 	echo "Usage:"
-	echo "  $(basename $0) [-f format] [-d directory] [-i index] [-t title] [-p processor]"
-	echo "  $(basename $0) -h"
+	echo "  $(basename "$0") [-f format] [-d directory] [-i index] [-t title] [-p processor]"
+	echo "  $(basename "$0") -h"
 	echo
 	echo "Optional arguments:"
 	echo "  -f output file format (either xhtml or html; default is $format)"
@@ -114,7 +114,6 @@ usage_help()
 	echo "  -p XSLT processor (xsltproc, xalan, or sabcmd; default is $processor)"
 	echo "  -h help"
 	echo
-	exit 1
 }
 
 create_index_file()
@@ -136,11 +135,11 @@ create_index_file()
 
 	echo "<head>" >> "$index_file"
 	echo "    <meta http-equiv=\"content-type\" content=\"text/html; charset=utf-8\"/>" >> "$index_file"
-	echo "    <title>"$index_title"</title>" >> "$index_file"
+	echo "    <title>$index_title</title>" >> "$index_file"
 	echo "    <link rel=\"stylesheet\" href=\"logtalk.css\" type=\"text/css\"/>" >> "$index_file"
 	echo "</head>" >> "$index_file"
 	echo "<body>" >> "$index_file"
-	echo "<h1>"$index_title"</h1>" >> "$index_file"
+	echo "<h1>$index_title</h1>" >> "$index_file"
 	echo "<ul>" >> "$index_file"
 
 	if [ -e "./directory_index.xml" ] ; then
@@ -149,16 +148,16 @@ create_index_file()
 		echo "    <li><a href=\"entity_index.html\">Entity index</a></li>" >> "$index_file"
 		echo "    <li><a href=\"predicate_index.html\">Predicate index</a></li>" >> "$index_file"
 	else
-		for file in $(grep -l "<logtalk_entity" *.xml); do
+		for file in $(grep -l "<logtalk_entity" ./*.xml); do
 			name="$(expr "$file" : '\(.*\)\.[^./]*$' \| "$file")"
 			entity=${name%_*}
 			pars=${name##*_}
 			echo "  indexing $name.html"
 			if [ $pars -gt 0 ]
 			then
-				echo "    <li><a href=\""$name".html\">"$entity"/"$pars"</a></li>" >> "$index_file"
+				echo "    <li><a href=\"$name.html\">$entity/$pars</a></li>" >> "$index_file"
 			else
-				echo "    <li><a href=\""$name".html\">"$entity"</a></li>" >> "$index_file"
+				echo "    <li><a href=\"$name.html\">$entity</a></li>" >> "$index_file"
 			fi
 		done
 	fi
@@ -167,7 +166,7 @@ create_index_file()
 
 	date="$(eval date)"
 
-	echo "<p>Generated on "$date"</p>" >> "$index_file"
+	echo "<p>Generated on $date</p>" >> "$index_file"
 	echo "</body>" >> "$index_file"
 	echo "</html>" >> "$index_file"
 }
@@ -180,13 +179,13 @@ do
 		i) i_arg="$OPTARG";;
 		t) t_arg="$OPTARG";;
 		p) p_arg="$OPTARG";;
-		h) usage_help;;
-		*) usage_help;;
+		h) usage_help; exit;;
+		*) usage_help; exit;;
 	esac
 done
 
 if [ "$f_arg" != "" ] && [ "$f_arg" != "xhtml" ] && [ "$f_arg" != "html" ] ; then
-	echo "Error! Unsupported output format: $f_arg"
+	echo "Error! Unsupported output format: $f_arg" >&2
 	usage_help
 	exit 1
 elif [ "$f_arg" != "" ] ; then
@@ -194,7 +193,7 @@ elif [ "$f_arg" != "" ] ; then
 fi
 
 if [ "$d_arg" != "" ] && [ ! -d "$d_arg" ] ; then
-	echo "Error! directory does not exists: $d_arg"
+	echo "Error! directory does not exists: $d_arg" >&2
 	usage_help
 	exit 1
 elif [ "$d_arg" != "" ] ; then
@@ -210,7 +209,7 @@ if [ "$t_arg" != "" ] ; then
 fi
 
 if [ "$p_arg" != "" ] && [ "$p_arg" != "xsltproc" ] && [ "$p_arg" != "xalan" ] && [ "$p_arg" != "sabcmd" ] ; then
-	echo "Error! Unsupported XSLT processor: $p_arg"
+	echo "Error! Unsupported XSLT processor: $p_arg" >&2
 	usage_help
 	exit 1
 elif [ "$p_arg" != "" ] ; then
@@ -249,10 +248,10 @@ if ! [ -e "$directory/logtalk.css" ] ; then
 	cp "$LOGTALKUSER"/tools/lgtdoc/xml/logtalk.css "$directory"
 fi
 
-if [ $((grep -l "<logtalk" *.xml | wc -l) 2> /dev/null) -gt 0 ] ; then
+if [ $( (grep -l "<logtalk" ./*.xml | wc -l) 2> /dev/null) -gt 0 ] ; then
 	echo
 	echo "converting XML files..."
-	for file in $(grep -l "<logtalk_entity" *.xml); do
+	for file in $(grep -l "<logtalk_entity" ./*.xml); do
 		echo "  converting $file"
 		name="$(expr "$file" : '\(.*\)\.[^./]*$' \| "$file")"
 		case "$processor" in
@@ -261,7 +260,7 @@ if [ $((grep -l "<logtalk" *.xml | wc -l) 2> /dev/null) -gt 0 ] ; then
 			sabcmd)		eval sabcmd \"$entity_xslt\" \"$file\" \"$directory\"/\"$name.html\";;
 		esac
 	done
-	for file in $(grep -l "<logtalk_index" *.xml); do
+	for file in $(grep -l "<logtalk_index" ./*.xml); do
 		echo "  converting $file"
 		name="$(expr "$file" : '\(.*\)\.[^./]*$' \| "$file")"
 		case "$processor" in