+++ tools/lgtdoc/xml/lgt2pdf.sh
@@ -3,7 +3,7 @@
 #############################################################################
 ## 
 ##   XML documenting files to PDF conversion script 
-##   Last updated on February 8, 2018
+##   Last updated on May 16, 2018
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -46,15 +46,15 @@ if ! [ "$LOGTALKHOME" ]; then
 		LOGTALKHOME="$( cd "$( dirname "$0" )" && pwd )/.."
 		echo "... using Logtalk installation found at $( cd "$( dirname "$0" )" && pwd )/.."
 	else
-		echo "... unable to locate Logtalk installation directory!"
+		echo "... unable to locate Logtalk installation directory!" >&2
 		echo
 		exit 1
 	fi
 	echo
 elif ! [ -d "$LOGTALKHOME" ]; then
-	echo "The environment variable LOGTALKHOME points to a non-existing directory!"
-	echo "Its current value is: $LOGTALKHOME"
-	echo "The variable must be set to your Logtalk installation directory!"
+	echo "The environment variable LOGTALKHOME points to a non-existing directory!" >&2
+	echo "Its current value is: $LOGTALKHOME" >&2
+	echo "The variable must be set to your Logtalk installation directory!" >&2
 	echo
 	exit 1
 fi
@@ -98,8 +98,8 @@ usage_help()
 	echo "current directory to PDF files"
 	echo
 	echo "Usage:"
-	echo "  $(basename $0) [-f format] [-d directory] [-p processor]"
-	echo "  $(basename $0) -h"
+	echo "  $(basename "$0") [-f format] [-d directory] [-p processor]"
+	echo "  $(basename "$0") -h"
 	echo
 	echo "Optional arguments:"
 	echo "  -f paper format (either a4 or us; default is $format)"
@@ -107,7 +107,6 @@ usage_help()
 	echo "  -p XSL-FO processor (either fop, xep, or xinc; default is $processor)"
 	echo "  -h help"
 	echo
-	exit 1
 }
 
 while getopts "f:d:p:h" Option
@@ -116,13 +115,13 @@ do
 		f) f_arg="$OPTARG";;
 		d) d_arg="$OPTARG";;
 		p) p_arg="$OPTARG";;
-		h) usage_help;;
-		*) usage_help;;
+		h) usage_help; exit;;
+		*) usage_help; exit;;
 	esac
 done
 
 if [ "$f_arg" != "" ] && [ "$f_arg" != "a4" ] && [ "$f_arg" != "us" ] ; then
-	echo "Error! Unsupported output format: $f_arg"
+	echo "Error! Unsupported output format: $f_arg" >&2
 	usage_help
 	exit 1
 elif [ "$f_arg" != "" ]
@@ -131,7 +130,7 @@ then
 fi
 
 if [ "$d_arg" != "" ] && [ ! -d "$d_arg" ] ; then
-	echo "Error! directory does not exists: $d_arg"
+	echo "Error! directory does not exists: $d_arg" >&2
 	usage_help
 	exit 1
 elif [ "$d_arg" != "" ] ; then
@@ -139,7 +138,7 @@ elif [ "$d_arg" != "" ] ; then
 fi
 
 if [ "$p_arg" != "" ] && [ "$p_arg" != "fop" ] && [ "$p_arg" != "xep" ] && [ "$p_arg" != "xinc" ] ; then
-	echo "Error! Unsupported XSL-FO processor: $p_arg"
+	echo "Error! Unsupported XSL-FO processor: $p_arg" >&2
 	usage_help
 	exit 1
 elif [ "$p_arg" != "" ] ; then
@@ -164,10 +163,10 @@ if ! [ -e "./logtalk_entity.xsd" ] ; then
 	cp "$LOGTALKHOME"/tools/lgtdoc/xml/logtalk_entity.xsd .
 fi
 
-if [ $((grep -l "<logtalk" *.xml | wc -l) 2> /dev/null) -gt 0 ] ; then
+if [ $( (grep -l "<logtalk" ./*.xml | wc -l) 2> /dev/null) -gt 0 ] ; then
 	echo
 	echo "converting XML files to PDF..."
-	for file in $(grep -l "<logtalk_entity" *.xml); do
+	for file in $(grep -l "<logtalk_entity" ./*.xml); do
 		echo "  converting $file"
 		name="$(expr "$file" : '\(.*\)\.[^./]*$' \| "$file")"
 		case $processor in